Move Config into crate root
